Add all the files needed for this program
This is the same program as the project from day 15, a virtual coffee machine. But in this version I learn object oriented programming and the use of object, class, method and attribute. I must say this is a lot of new information for me and I am a bit confused. But I understand the logic of my code. I just need more practice.